First version that compiled successfully
First version that compiled successfully, most likely still does not
work.

It copies all the data to the GPU and then works with it. This version
also consolidates the medoids to continuous GPU global memory, so that
it is easier to read. This may be slower than just transfering medoid
indexes and reading them from the whole dataset in GPU memory, but it
was already implemented to run it on dataset of any size, which requires
this.

Basically the dataset is split up so that each MPI rank receives a chunk
of the same size. This dataset is called nodeData, or source data, in
the code.

Then for assignment, we consolidate medoids into continuous GPU memory.
We split the nodeData to small subsets where each is processed by a
single kernel. These subsets are then further divided into smaller
subsets, small enough that they fit into shared memory of the block.
Then each block reads the whole nodeData subset assigned to it into
hsared memory and goes through all the medoids, each time reading it
into shared memory and then calculating SQFD between medoid and data
signature.

The whole block calculates the given SQFD, so that the matrix is
calculated fast.

For each block data signature we keep min distance and current
assignment in shared memory, which is only updated by thread 0 to make
synchronization easier. After the block finishes it's assigned
calculation, the data signatures assigned to it should have their
assignment finished.

The assignments are then shared with the rest of the cluster using
MPI_Allgather, which does exactly what we need. It eansures that all
cluster nodes have the whole assignment data.

The assignment data is required to calculate new medoids. This is done
in O(n^2) complexity to maximize the available parallelism and minimize
the need to move data around in memory. This is the place where we might
need to optimize.

We basically do the same thing as in assignment kernel, but in this case
we take the whole data set as distance targets. So each block receives a
subset of node data and calculates the sum of distance to each other
signature in the same cluster. This is done by going through the whole
dataset.

As distance between each pair is calculated by a single GPU block, there
is no code diversion, it just skips the calculation.
The only downside is that each block of each kernel must go through the
whole dataset, which may push on the data path between global memory and
SMPs.

When scores of each signature are calculated, we again exchange them
between all nodes of the cluster using MPI_Allgather. This ensures that
the whole cluster has synchronized idea about the clusterization for the
next iteration.

From the scores, each node separately calculates the medoids for the
next iteration. This is just linear pass through the scores and
assignments and finding minimum score for each cluster. This part has
too much data transfer to processing needed to make it viable as GPU
calculation. Due to this, we should use OpenMP to parallellize this
calculation. It is basically a reduce operation.

If no medoid changed, we finish the iteration.