First working version, applied fix from Slack to sqdf.hpp, changed compilation to debug mode
Changed Makefile in parallel to compile both the CUDA code and the C++
code with symbols and without optimizations, which aids in debugging.

Added paralell/run.sh file which runs the parallel version using SLURM.
It's basically a sbatch submit script without using sbatch.

Added parallel/parallel.sh, which runs the parallel solution using
cuda-memcheck to check for any invalid manipulation of memory.

Added per node redirection of stdout and stderr to file, which redirects
the given stream to a rankN.out and rankN.err respectively. This is so
that it is not interleaved and is easier to debug.

Added sorting of initial medoids, as that is what the serial solution
does and I need to check my results against serial results. But seems
like unnecessary work, so will remove in the finished implementation.

Fixed serial Makefile, which was setup for the source directory under
/mnt/home/_teaching/advpara/final-kmedoids or something. It did not take
into account that the symlink they have in the directory is not commit
into git and will not work when the directory is cloned or copied.

Fixed both assignment and score kernel implementation. There were many
errors, like off by one errors, forgotten DIM and DIM + 1
multiplications etc.

The most difficult to debug were switched parameters of cudaMemcpy,
which takes DESTINATION first, then source. Also wrong type when
calculating size to copy, for example when moving assignments onto the
GPU in Score kernel, I mistakenly had FLOAT instead of size_t.

Also switched to Allgatherv, so that we can handle the situation where
the last node has less data to process easily. I tried to just allocate
slightly larger arrays, but it then messes up with rest of the program,
as it uses the size of the vectors to get the number of images in the
dataset etc.

Fixed missing DIM + 1 multiplication in shared memory size computation.
Before, we were calculating the size in number of centroids instead of
number of floats.

Added redirecter and other helpers to helpers.hpp, which help me
redirect file descriptors, C++ streams and to print vectors.

Improved mergeSums device functions, as it is used in many places in the
CUDA code and might be a bottleneck. The current implementatin is very
fast, I don't think I can get it any faster without some serious hacks.

Added weigth normalization, which I completely missed that it was
necessary. Weights are normalized so that they sum up to 1. It does not
matter if they are different for each signature, this makes them work
together well.

Found out that accessing any memory, including shared memory, MUST be
done at multiple of the size of the accessed datatype. So pointer to
float must be pointing to an address divisible by 4 and pointer to
db_offset_t must be divisible by 8. This means that if i put some FLOAT
array in shared memory before the db_offset_array and it has odd number
of elements, then accessing the db_offset_t array will fail. This forces
us to order the shared memory arrays by data type size in decreasing
order, so first the 8 byte types, then the 4 byte types and then
smaller. Also the shared memory declaration has to be using the largest
datatype, so that it is aligned properly.

The hardest error to spot was the one when computing sigPerBlock. I am
using the ?: operator, which has lowest precedence, which I did not
know. So everything, even the numSig / gridDim.x, was part of the
condition, which of course did not work. So we added parantheses to fix
this.

Added some basic filtering of the data processed when computing the
scores, but most of it is commented out as it will be rewritten
completely.