Reimplemented the CUDA algorithm
This commit contains the first version that compiled successfully, which
will most likely be broken and full of bugs. The next commit will fix
that.

Reimplemented the cuda algorithm, creating new MemAlg. This algorithm
takes advantage of the fact that the whole dataset fits into device
memory. This means that we upload the data once and then just upload
offsets to this data.

This goes mostly for assignments. Before, each assignment computation
required us to pull medoids into continuous memory on host and then
transfer this data to device. The problem is the data already was on the
device, just not continuous. But due to the size of eachsignature, which
is up to 144*8*4 Bytes long, we really don't need the medoids to be
continuous as they are large enough that it makes no difference.

So now we only pass indexes of the medoids and the kernel reads them
from the whole dataset.

As for score computation, I came up with a hopefully better algorithm.
Before, we used the straightforward division of dataset and computed the
score for each source image in the dataset for given rank. During this
computation, we had to go through the whole dataset in each kernel
block, making it really slow.

In the new algorithm, we compute a clusterList, which is a list of
images in order of their assignment to clusters. In other words, first
we have indexes of all images in cluster 0, then all in cluster 1, then
cluster 2 etc. To suplement this data structure, we also have
clusterListIndex, which tells us index into the clusterList where the
given cluster starts. So clusterList[clusterListIndex[i]] is the first
image belonging to the cluster i.

This datastructure allows us to enumerate all images from a given
cluster without the need to go through the whole dataset. Even though
the image signatures are not continuous in memory, due to their size as
we discussed above with the medoids, this is not a problem.

To use this datastructure, I added two new kernels. One is optimized to
process part of a large cluster. This kernel allows us to compute the
score of a given source image by multiple blocks or kernels, splitting
up the target range into multiple kernels.

The second kernel is optimized to process a sequence of small clusters,
where each block may process images from multiple clusters.

Both kernels use the same structure as the original kernel. Each block
loads a given small number of source signatures to shared memory and
then computes their score, compcomputing the SQFD to all other images in
the given cluster.

In the big cluster kernel, multiple blocks may load the same source
signatures to shared memory and compute their distance to differnet
parts of the cluster, splitting up the cluster into multiple target
ranges, increasing parallelism. Once they finish processing their target
range, these blocks use atomic operations to add the score to the
result.

The small cluster kernel is little more complex. It also loads source
signatures to shared memory, but these signatures may be from different
clusters. The clusters will be continuous, so for example clusters 1,2,3
or 7,8,9 etc. The target range then becomes the whole clusters from
which the source signatures are loaded, which creates a continuous part
of clusterList array. We then go through this target range sequentialy,
for each target signature going only through the source signatures
belonging to the same cluster. This way we calculate the whole score of
each source signature in single kernel block, which we then just set as
the value in the sScores.

During the implementation of the new algorithm I also decided to improve
the structure of the repository. Now the base algorithm class as well as
each algorithm implementation are in separate files, making it more
readable. The base algorithm is in parallel/headers/alg.hpp, the old
algorithm in parallel/headers/consAlg.hpp and the new algorithm in
parallel/headers/memAlg.hpp.

Removed the work.hpp file, moving the main algorithm method into the
algorithm class itself. Each algorithm now implements constructor,
initialize, runIteration, destructor and fillResults/moveResults
methods. We first construct the instance, do more initialization, then
run the required number of iterations while runTIterations returns true,
then retrieve results and the destroy the instance.