Many many bugfixes, should be working
Fixed source data consolidation in both small and large kernel. In the
copySigData call, there was shData instead of cshData, which ment that
we were copying all signatures onto the first signature, overwriting it
and leaving the rest uninitialized.

Added shScores zeroing at the start of both kernels

Fixed the output scores copying in large kernel, which was mistakenly
merging outputs from all blocks where the outputs of the first block
should go.

Added check so that blocks which have no source signatures to process
just end.

Fixed lastClusterEnd calculation. We were using clusterListIndex in
assignment array, while also getting the cluster after the next cluster.

Fixed output scores of small kernel. Used wrong for cycle, which ment
that all threads were reading and writing the same values, coliding
everywhere.

For now, I have left all debug outputs in the code, just commented out.
I should remove them before submission so that noone sees the way I
debug things.

Fixed clusterListIndex computation. We now allocate numCLusters+2. +1 so
that we can have leading 0 for easier computations, and another +1 so
that we can use the hClusterListIndex to fill hCLusterList. When filling
hClusterList, we add to each item in hClusterLIstIndex when we add
another image into a cluster. This way we track the where the next image
in the given cluster goes in the hClusterList array. The result is that
the whole hCLusterListINdex array is shifted one item to the right, i.e.
towards the lower indexes. When copying the array to the device, to the
dClusterListIndexes, we then leave out the last item, as it is just a
copy of the one before last.

Fixed asgnProps, which were getting inputData.dataSize instead of size,
which gave us invalid rankDataSize, which made first rank do all the
work. It also had uninitialized sigPerBlock and blockSize.

Added print methods to both asgnProps and scoreProps for easier
debugging.

Fixed allgatherv calls. These calls MUST not have overlapping sendbuf
and recvbuf, which I had. For this, they have a special constant
MPI_IN_PLACE, which we can pass instead of the sendbuf argument and it
expects the data which would be passed to sendbuf to already be present
in the correct place in the recvbuf, from where it will take it to send
it to rest of the ranks.

Added missing device uploads of assignments, clusterListIndex and
clusterList. These are done in a new dataStream CUDA stream, which is
dedicated to data transfer and for now is only used during CPU
computations of clusterList, clusterListIndex etc.

Added dScores zeroing, which zeroes out the dScores array while the
clsuterList, clusterListIndex are getting transfered.

Fixed clusterListIndex computation, which did not take into account the
leading 0 and the right shift.

Added clusterComplexities computation.

Fixed scores computation. The search for the first cluster to be
processed by the current rank was just broken. binary_search returns
bool, which somehow the compiler did not mind converting to std::size_t,
and so it broke everything. Now we find the first cluster which will be
surely processed by the current rank, which means it's complexity is
greater than the startComplexity of the current rank. We then move
curCluster to first smaller than the startComplexity, which is the first
cluster which may be processed by the current rank, depending if it is a
lrage cluster and it was not processed in its entirety by the previous
rank.

Instead of using iallgather, used Iallgatherv. I don't know why, but
allgather just was not working. It just sent the data of the rank 0 to
everyone, but left everything else zeroed out. Allgatherv works well.

Fixed broken diplacement computation, which made us send wrong data when
sharing rankFragments

Fixed the partTOStartAt and partToEndAt computation. It was using
integer division instead of float division, which lead to bad results,
where we did not compute scores of some parts or computed scores of
other parts twice.

Renamed misleading blockStartSource etc. to kernelStartSource etc. which
better tells us what the variables mean. Block is just an overloaded
designation.

Move kernelSourceNUm computation inside the while cycle, as before we
just left the value of the first kernel for all kernels, which is wrong.

Fixed streamIdx incrementation, which was not looped around and was
accessing out of bound of the streams array.

Added missing numBlocks computation to small blocks handling. Before, we
were just starting max number of blocks which mostly were doing nothing.

Added deallocation of device and host memory.