Fixed bugs, optimized default argument values
Ran few measurement runs, measuring the program behavior when run with
different argument values. Some of the results can be seen in the
score_measurements.csv files. From these, I just took the fastest
setting and set that as the default value.

Added separate blockSize for Score computation. Before, it was using the
base blockSize, which was shared with Assignment computation.

Added smallClusterComplexity as program argument instead of a compile
time constant.

Used bpplib and its stopwatch to measure each separate part of the
algorithm, namely Assignment, preprocessing and scores, and used this
for the optimization of default argument values.

Fixed invalid receive counts and displacements in initialization, as
they were using rankDataSize, which is different for the last rank and
thus the last rank was providing invalid displacements and recvCounts.

Changed processBigCluster and processSmallCluster methods so that they
take the ScoreProperties class instead of disassembling the class to the
values it holds.

Fixed partTOStartAt and partToEndAt computation. Before, they were not
manually typecast to std::size_t from double, so I have no idea how that
worked. Now we properly type cast them and also clamp them so that start
is not before the cluster start and end is not after the cluster end,
while also properly rounding them, so that they overlap. It does not
really matter if they overlap, as both ranks should then compute the
same value for the overlapping image and if it a minimum, it should be
chosen by both nodes and then by all nodes when processing fragments.

Fixed targetsNum in processBigCluster, which was outside the while loop
and was not changed in sync with targetsStart and targetsEnd. This lead
to out of bound memory access in the gpu, as the  kernel was given
targetsNum larger than rest of the cluster.